added custom encodings

+++ misc/csvUtils.py
@@ -1,10 +1,7 @@
 import csv
 import os
 import pandas as pd
-from progress.bar import IncrementalBar
 from typing import List
-import geneticAlgorithm.encoding as encoding
-import geneticAlgorithm.fitnessFunctions as functions
 
 def mergeFreqs(fileNames, fitnessFunction, outputFile):
     """
@@ -65,10 +62,24 @@ def updateCSV(inputPath, data=None, headers = None, overwrite=False):
     '''
     Takes in an input path and a 2D list, then adds the contents of the list to the given input file
     '''
-    if not os.path.exists(inputPath) or overwrite:
+    dirs = inputPath.split('/')
+
+    for i in range(len(dirs)):
+        if dirs[i] == '.':
+            continue
+
+        if i == len(dirs) - 1:
+            break
+        
+        path = "/".join(dirs[:i + 1])
+        if not os.path.exists(path):
+            os.mkdir(path)
+
+    if (not os.path.exists(inputPath) and headers) or overwrite:
         with open(inputPath, 'w+', newline='') as out:
             writer = csv.writer(out)
-            writer.writerow(headers)
+            if headers:
+                writer.writerow(headers)
             out.close()
 
     if data:
